Provide access to header to model conversions.
